change H2Config, merge H2Config and H2TestConfig java file, H2TestConfig is unnecessary because we can use the Profile annotation
