+++ patient-dao/src/test/java/hu/kuncystem/patient/dao/user/JDBCUserDaoTest.java
@@ -9,10 +9,11 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
-import hu.kuncystem.patient.dao.H2TestConfig;
+import hu.kuncystem.patient.dao.H2Config;
 import hu.kuncystem.patient.pojo.user.User;
 import hu.kuncystem.patient.pojo.user.UserFactory;
 import hu.kuncystem.patient.pojo.user.UserGroup;
@@ -29,7 +30,8 @@ import static org.hamcrest.Matchers.instanceOf;
  * @version 1.0
  */
 @RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = { H2TestConfig.class })
+@ContextConfiguration(classes = { H2Config.class })
+@ActiveProfiles("test")
 public class JDBCUserDaoTest {
 
     @Autowired
